requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
python-multipart
pydantic
pydantic-settings
alembic
jinja2
reportlab
httpx
pytest
